// frontend/src/app/api/model/bookCreateDto.ts
/**
 * Book Shelf API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BookCreateDto {
  title: string;
  isbn?: string;
  pages: number;
  yearPublished: number;
  summary: string;
  format: BookCreateDto.FormatEnum;
  authorId: string;
  languageId: string;
  publisherId: string;
  genreIds: Set<string>;
}

export namespace BookCreateDto {
  export const FormatEnum = {
    Hardcover: 'HARDCOVER',
    Paperback: 'PAPERBACK',
    Digital: 'DIGITAL'
  } as const;
  export type FormatEnum = typeof FormatEnum[keyof typeof FormatEnum];
}